move the check about unsupported literals directly to the parser
